fix: delete user in pdv

+++ src/pdv/pdv.service.ts
@@ -5,6 +5,7 @@ import { Pdv } from './pdv.entity';
 
 @Injectable()
 export class PdvService {
+  [x: string]: any;
   constructor(
     @InjectRepository(Pdv)
     private pdvRepository: Repository<Pdv>,
@@ -32,6 +33,14 @@ export class PdvService {
   }
 
   async remove(id: number): Promise<void> {
+    const pdv = await this.pdvRepository.findOne({
+      where: { id },
+    });
+
+    if (!pdv) {
+      throw new Error(`PDV com id ${id} não encontrado`);
+    }
+
     await this.pdvRepository.delete(id);
   }
 }
